@@ -4,6 +4,8 @@ import babel from 'rollup-plugin-babel';
 import json from 'rollup-plugin-json';
 import commonjs from 'rollup-plugin-commonjs';
 
+import {dependencies, peerDependencies} from '../../package.json';
+
 import styles from './plugins/styles';
 import image from './plugins/image';
 import icon from './plugins/icon';
@@ -14,23 +16,13 @@ const project = resolve(__dirname, '../..');
 const buildRoot = resolve(project, './build-intermediate');
 const styleRoot = resolve(buildRoot, './styles');
 
-const PACKAGES = [
-  'react',
-  'react-dom',
-  'babel-runtime',
-  'tslib',
-  'lodash',
-  'core-js',
-  'hoist-non-react-statics',
-  '@shopify/javascript-utilities',
-  '@shopify/react-utilities',
-];
+const externalPackages = [...Object.keys(dependencies), ...Object.keys(peerDependencies)];
 
 export default function createRollupConfig({entry, outputCSS}) {
   return {
     entry,
     external(id) {
-      return PACKAGES.some((aPackage) => id.startsWith(aPackage));
+      return externalPackages.some((aPackage) => id.startsWith(aPackage));
     },
     plugins: [
       json(),
