@@ -1,5 +1,5 @@
 import EASDK, {Options} from './EASDK';
-import {withEASDK, WithEASDKProps} from './decorator';
+import {withEASDK, WithEASDKProps, contextTypes} from './decorator';
 
 export default EASDK;
-export {Options, withEASDK, WithEASDKProps};
+export {Options, withEASDK, WithEASDKProps, contextTypes};
